Move user info to global state
