+++ packages/nextjs/pages/_app.tsx
@@ -5,7 +5,7 @@ import "@rainbow-me/rainbowkit/styles.css";
 import NextNProgress from "nextjs-progressbar";
 import { Toaster } from "react-hot-toast";
 import { useDarkMode } from "usehooks-ts";
-import { WagmiConfig } from "wagmi";
+import { WagmiConfig, useAccount } from "wagmi";
 import { Footer } from "~~/components/Footer";
 import { Header } from "~~/components/Header";
 import { BlockieAvatar } from "~~/components/scaffold-eth";
@@ -13,14 +13,17 @@ import { useNativeCurrencyPrice } from "~~/hooks/scaffold-eth";
 import { useGlobalState } from "~~/services/store/store";
 import { wagmiConfig } from "~~/services/web3/wagmiConfig";
 import { appChains } from "~~/services/web3/wagmiConnectors";
+import { web3AuthInstance } from "~~/services/web3/wagmiConnectors";
 import "~~/styles/globals.css";
 
 const ScaffoldEthApp = ({ Component, pageProps }: AppProps) => {
   const price = useNativeCurrencyPrice();
   const setNativeCurrencyPrice = useGlobalState(state => state.setNativeCurrencyPrice);
+  const setUserInfo = useGlobalState(state => state.setUserInfo);
   // This variable is required for initial client side rendering of correct theme for RainbowKit
   const [isDarkTheme, setIsDarkTheme] = useState(true);
   const { isDarkMode } = useDarkMode();
+  const { connector } = useAccount();
 
   useEffect(() => {
     if (price > 0) {
@@ -31,6 +34,21 @@ const ScaffoldEthApp = ({ Component, pageProps }: AppProps) => {
   useEffect(() => {
     setIsDarkTheme(isDarkMode);
   }, [isDarkMode]);
+
+  useEffect(() => {
+    const getUserInfo = async () => {
+      try {
+        if (web3AuthInstance) {
+          const userInfo = await web3AuthInstance.getUserInfo();
+          console.log(userInfo);
+          setUserInfo(userInfo);
+        }
+      } catch (error) {
+        console.error(error);
+      }
+    };
+    getUserInfo();
+  }, [connector]);
 
   return (
     <WagmiConfig config={wagmiConfig}>
